% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseball}
\alias{baseball}
\title{baseball Data Set}
\format{
A tibble with 337 rows and 28 variables:
\describe{
\item{salary}{The salary in $1000s.}
\item{average}{Batting average of the player.}
\item{obp}{On base percentage of the player}
\item{runs}{Number of runs scored.}
\item{hits}{Number of hits in total.}
\item{doubles}{Number of doubles hit.}
\item{triples}{Number of triples hit.}
\item{homeruns}{Number of homeruns hit.}
\item{rbis}{Number of runs batted in.}
\item{walks}{Number of times walked.}
\item{sos}{Number of strikeouts.}
\item{sbs}{Number of stolen bases.}
\item{errors}{Number of errors committed.}
\item{freeagent}{Factor indicating whether the player is a free agent or
is eligible for free agency.}
\item{arbitration}{Factor indicating whether the player has arbitration or
is eligible for arbitration.}
\item{runsperso}{Number of runs per strikeout (runs/sos).}
\item{hitsperso}{Number of hits per strikeout (hits/sos).}
\item{hrsperso}{Number of homeruns per strikeout (homeruns/sos).}
\item{rbisperso}{Number of rbis per strikeout (rbis/sos).}
\item{walksperso}{Number of walks per strikeout (walks/sos).}
\item{obppererror}{On base percentage per error (obp/errors).}
\item{runspererror}{Number of runs scored per error (runs/errors).}
\item{hitspererror}{Number of hits per error (hits/errors).}
\item{hrspererror}{Number of homeruns per error (homeruns/errors).}
\item{sospererror}{Number of strikeouts per error (sos/errors).}
\item{sbsobp}{Number of stolen bases times on base percentage (sbs\emph{obp).}
\item{sbsruns}{Number of stolen bases times number of runs scored
(sbs}runs).}
\item{sbshits}{Number of stolen bases times number of hits (sbs*hits).}
}
}
\usage{
baseball
}
\description{
This data set contains information for 337 baseball players.
}
\keyword{datasets}
