% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births}
\alias{births}
\title{births Data Set}
\format{
A tibble with 278 rows and 2 variables:
\describe{
\item{prenatal}{A factor indicating the prenatal care the mother received:
Adequate, Inadequate, or Intensive.}
\item{type}{A factor indicating the classification of the birth: "Preterm
(induced or cesarean)", "Preterm (without procedures)", and
"Term or post-term"}
}
}
\usage{
births
}
\description{
Data from 1995-1997 for a study hat examined pregnancies that resulted in the
birth of twins. Births were classified as preterm with intervention (induced
labor or cesarean), preterm without procedures, or term/post-term.
Researchers also classified the pregnancies by the level of prenatal medical
care the mother received (inadequate, adequate, or intensive). The data set
consists of 278 cases (rows) with two columns indicating the level of
prenatal care and type of birth for each set of twins.
}
\keyword{datasets}
