% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Outputs and restyles an object from the \code{boot.ci()} function from the boot
package.}
\usage{
boot_ci(boot_out, conf = 0.95, type = "all", ...)
}
\arguments{
\item{boot_out}{An object of class "boot" or "boot3150" containing the output
of a bootstrap calculation.}

\item{conf}{A scalar or vector containing the confidence level(s) of the
required interval(s).}

\item{type}{A vector of character strings representing the type of intervals
required. The value should be any subset of the values c("norm","basic",
"stud", "perc", "bca") or simply "all" which will compute all five types of
intervals.}

\item{...}{See more documentation in the \link[boot]{boot.ci} function.}
}
\value{
This function returns an object of class boot3150ci that works much
like an object of class bootci from the \code{boot} package.
}
\description{
This function relies on the \link[boot]{boot.ci} function. This function
generates 5 different types of equi-tailed two-sided nonparametric
confidence intervals. These are the first order normal approximation, the
basic bootstrap interval, the studentized bootstrap interval, the bootstrap
percentile interval, and the adjusted bootstrap percentile (BCa) interval.
All or a subset of these intervals can be generated.
}
\examples{
med <- function(x, i){median(x[i])}
sample_data <- rnorm(100)
boot_out <- bootstrapping(sample_data, med, 1000)
bootci_out <- boot_ci(boot_out)

}
