% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cars99}
\alias{cars99}
\title{cars99 Data Set}
\format{
A tibble with 109 rows and 11 variables:
\describe{
\item{Model}{The vechicle model name.}
\item{CityMPG}{The miles per gallon (MPG) for the vehicle in the city.}
\item{HwyMPG}{The miles per gallon (MPG) for the vehicle on the highway}
\item{FuelCap}{The fuel capacity of the vehicle in gallons.}
\item{Weight}{The weight of the vehicle in lbs.}
\item{FrontWt}{The front weight of the vehicle.}
\item{Accel0_30}{The time it takes, in seconds, for the vehicle to
accelerate from 0 to 30 mph.}
\item{Accel0_60}{The time it takes, in seconds, for the vehicle to
accelerate from 0 to 60 mph.}
\item{QtrMile}{The time it takes, in seconds, for the vehicle to travel a
quarter of a mile.}
}
}
\usage{
cars99
}
\description{
This data set contains information on 109 vechicles from 1999.
}
\keyword{datasets}
