% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_data_f2019}
\alias{class_data_f2019}
\title{class_data_f2019 Data Set}
\format{
A tibble with 42 rows and 7 variables:
\describe{
\item{level}{A factor indicating the class level the student is: Freshman,
Sophomore, Junior, Senior, or Graduate.}
\item{major}{A character indicating the major of the student.}
\item{sex}{A factor indicating the sex of the student: F for female or M
for male.}
\item{ski}{A factor indicating downhill preference: Ski, Snowboard, or
Neither.}
\item{penny}{A factor indicating preference regarding the penny: Abolish,
Retain or No Answer.}
\item{speed}{An integer indicating the fastest speed the student had driven
in a vehicle (in mph).}
\item{sleep}{A numeric variable indicating how long the student slept the
night before.}
}
}
\usage{
class_data_f2019
}
\description{
This data set contains information on 42 students from Fall of 2019.
}
\keyword{datasets}
