% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_plot.R
\name{classifier_plot}
\alias{classifier_plot}
\title{Plots the Points and Predicted Region for a Classifier Model}
\usage{
classifier_plot(classifier_model, data, res = 200, sv_diamonds = TRUE)
}
\arguments{
\item{classifier_model}{The classifier model of class "naive_bayes", "lda",
or "svm".}

\item{data}{The dataset that should be plotted. This need not be the dataset
on which the classifier was trained.}

\item{res}{The resolution of the plot. Larger values will make the lines look
straighter and make the grid more fine, but take longer.}

\item{sv_diamonds}{A boolean indicating whether the support vectors should be
plotted with diamonds around them when plotting a support vector machine
classifier.}
}
\value{
This function returns a ggplot object plot. It can be added onto like
any ggplot object.
}
\description{
This function plots the points and predicted region for a classifier model of
class "naive_bayes", "lda", or "svm". This function can support up to three
predictor variables: two of which can be quantitative and one can be
categorical.
}
\examples{
data(mtcars)
mtcars2 <- mtcars |> mutate(vs = factor(vs), am = factor(am))
classifier <- MASS::lda(vs ~ mpg + disp + am, data = mtcars2)
classifier_plot(classifier, data = mtcars2)
plot_tukey(t)

}
