% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diseases}
\alias{diseases}
\title{diseases Data Set}
\format{
A tibble with 51 rows and 4 variables:
\describe{
\item{State}{A charcter vector indicating the state.}
\item{AIDS}{The number of reporeted AIDS cases.}
\item{Syphilis}{The number of reporeted syphilis cases.}
\item{Tuberculosis}{The number of reporeted tuberculosis cases.}
}
}
\usage{
diseases
}
\description{
This data set contains information for each state and Washington, D.C. about
the number of reported cases of AIDS, syphilis, and tuberculosis.
}
\keyword{datasets}
