% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epilepsy}
\alias{epilepsy}
\title{epilepsy Data Set}
\format{
A tibble with 59 rows and 4 variables:
\describe{
\item{id}{The patient ID.}
\item{numseiz}{The number of seizures the patient had.}
\item{age}{The age of the patient}
}
}
\usage{
epilepsy
}
\description{
This data set contains information on the number of seizures, which
treatment, and the age of 59 patients with epilepsy.
}
\keyword{datasets}
