% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESL.mixture}
\alias{ESL.mixture}
\title{ESL Mixture Data Set}
\format{
A list with 8 components The components are
\describe{
\item{x}{200 x 2 matrix of training predictors.}
\item{y}{ class variable; logical vector of TRUES and FALSES - 100 of each.}
\item{xnew}{matrix 6831 x 2 of lattice points in predictor space.}
\item{prob}{vector of 6831 probabilities (of class TRUE) at each lattice point.}
\item{marginal}{marginal probability at each lattice point.}
\item{px1}{69 lattice coordinates for x.1.}
\item{px2}{99 lattice values for x.2  (69*99=6831).}
\item{means}{20 x 2 matrix of the mixture centers, first ten for one class, next ten for the other}
}
}
\usage{
ESL.mixture
}
\description{
These are the data used in Figures 2.1-2.3, and elsewhere through the
Elements of Statistical Learning book.
}
\keyword{datasets}
