% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idealwt}
\alias{idealwt}
\title{idealwt Data Set}
\format{
A tibble with 182 rows and 4 variables:
\describe{
\item{sex}{A factor indicating the sex of the person: Female or Male.}
\item{actual}{The person's actual weight.}
\item{ideal}{The person's ideal weight.}
\item{diff}{The difference between the person's actual weight and their
ideal weight. Negative values indicate that the person weighs less than
what they consider ideal.}
}
}
\usage{
idealwt
}
\description{
This data set contains weight information on 182 people (119 females and 63
males). Actual weights, ideal weights, and the difference between them are
recorded.
}
\keyword{datasets}
