% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilogit.R
\name{ilogit}
\alias{ilogit}
\title{Compute the Inverse Logit of a Value}
\usage{
ilogit(x)
}
\arguments{
\item{x}{A value for which that we want to find the inverse.}
}
\description{
A function that computes the inverse logit (the inverse log odds) of a value.
}
\examples{
data(mtcars)
mod <- glm(vs ~ mpg, data = mtcars, family = binomial())
ilogit(predict(mod, data.frame(mpg = 20)))

}
