% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_plots.R
\name{influence_plots}
\alias{influence_plots}
\title{Creates many diagnostic plots}
\usage{
influence_plots(model, missing_group = NULL)
}
\arguments{
\item{model}{The model for which we would like these plots. This can be of
class "lm" or "glm" with a binomial family.}

\item{missing_group}{Used for multinomial regression to indicate which group
is not being plotted.}
}
\value{
This function returns plots. The user must press "enter" or "return"
to view subsequent plots.
}
\description{
This function creates many diagnostic plots for a given model. These plots
include residual plots, a QQ plot, a leverage plot, a Cook's distance plot,
a DfFits plot, and DfBetas plots for the intercept and all slopes.
}
\examples{
mod <- lm(mpg ~ disp, data = mtcars)
influence_plots(mod)

}
