% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_plot.R
\name{interaction_plot}
\alias{interaction_plot}
\title{Two-way Interaction Plot with ggplot2}
\usage{
interaction_plot(
  df,
  response,
  x_factor,
  group_factor,
  linewidth = 1,
  point_size = 2,
  .f = mean,
  ...
)
}
\arguments{
\item{df}{The data frame that contains the response and factor variables.}

\item{response}{The name of the numeric variable giving the response. Can be
entered with or without quotes.}

\item{x_factor}{The name of a factor whose levels will form the x axis. Can
be entered with or without quotes.}

\item{group_factor}{The name of a factor whose levels will split up the other
variables. Can be entered with or without quotes.}

\item{linewidth}{The linewidth of the lines plotted. Enter 0 if the lines
should be omitted.}

\item{point_size}{The size of the points plotted. Enter 0 if the points
should be omitted.}

\item{...}{Other parameters that can be passed to \link[ggplot2]{geom_point}.}
}
\value{
This function returns a two-way interaction plot as a ggplot object.
All the ggplot add-on functions can be added onto the object returned for
more customization.
}
\description{
This function is similar to \link[stats]{interaction.plot}, but plots in
ggplot2 instead of base R. It plots the mean (or other summary) of the
response for two-way combinations of factors, thereby illustrating possible
interactions.
}
\examples{
data(warpbreaks)
interaction_plot(warpbreaks, breaks, wool, tension)
interaction_plot(warpbreaks, "breaks", "wool", "tension")

}
