% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_plots.R
\name{logistic_plots}
\alias{logistic_plots}
\title{Creates many diagnostic plots for logistic or multinomial models}
\usage{
logistic_plots(model)
}
\arguments{
\item{model}{The model for which we would like these plots. This can be of
class "glm" with a binomial family or of class "multinom" from the \code{nnet}
library.}

\item{missing_group}{Used for multinomial regression to indicate which group
is not being plotted.}
}
\value{
This function returns plots. The user must press "enter" or "return"
to view subsequent plots.
}
\description{
This function feeds into the \code{influence_plots()} function to create many
diagnostic plots for a given logistic or multinomial model. These plots
include residual plots, a leverage plot, a Cook's distance plot, a DfFits
plot, and DfBetas plots for the intercept and all slopes.
}
\examples{
mod <- glm(vs ~ disp, data = mtcars, family = "binomial")
logistic_plots(mod)

}
