% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlbsalaries}
\alias{mlbsalaries}
\title{mlbsalaries Data Set}
\format{
A tibble with 877 rows and 5 variables:
\describe{
\item{Rank}{The ranking of the player's salary with 1 being the most money
earned.}
\item{Name}{A character vector containing the name of the player.}
\item{Team}{A character vector containing the team the player played for.}
\item{Position}{A factor indicating what position the player played: SP for
starting pitcher, 1B for first base, 2B for second base, 3B for third base,
SS for shortstop, OF for outfield, RP for relief pitcher, and DH for
designated hitter.}
\item{Salary}{A numeric vector containing the salary the player made for
the 2018 season.}
}
}
\usage{
mlbsalaries
}
\description{
This data set contains information on 877 MLB players from 2018.
}
\keyword{datasets}
