% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nitrogen}
\alias{nitrogen}
\title{nitrogen Data Set}
\format{
A data frame with 36 rows 1 variable:
\describe{
\item{nitrogen}{The nitrogen content for the trees.}
}
}
\usage{
nitrogen
}
\description{
Nitrogen content of trees in an orchard, the growing tips of 150 leaves are
clipped from trees throughout the orchard. These leaves are ground to form
one composite sample, which the researcher assays for percentage of nitrogen.
Composite samples obtained from a random sample of 36 orchards throughout the
state gave the nitrogen contents.
}
\keyword{datasets}
