% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boot3150.R
\name{plot.boot3150}
\alias{plot.boot3150}
\title{Plots a boot3150 object}
\usage{
\method{plot}{boot3150}(x)
}
\arguments{
\item{x}{An object of class boot3150}
}
\value{
This function plots the boot3150 object.
}
\description{
When given a boot3150 object, this function plots the samples with both
histograms and QQ plots.
}
