% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tukey.R
\name{plot_tukey}
\alias{plot_tukey}
\title{Plots the Tukey HSD Confidence Intervals}
\usage{
plot_tukey(t, which = "all")
}
\arguments{
\item{t}{The Tukey HSD tibble outputted by the
\link[math3150package]{tukeyhsd} function.}

\item{which}{Indicates which variable for which the graph should be created.
The default is "all", but a variable name used in the model can also be
entered (in quotes).}
}
\value{
This function returns plots. The user must press "enter" or "return"
to view subsequent plots if there are any.
}
\description{
This function plots the Tukey HSD confidence intervals when given a tibble
outputted by the \link[math3150package]{tukeyhsd} function. If more than one
variable, all will be plotted by default, but that can be adjusted with the
\code{which} option.
}
\examples{
data(warpbreaks)
fm1 <- aov(breaks ~ wool + tension, data = warpbreaks)
t <- tukeyhsd(fm1, "tension", ordered = TRUE)
plot_tukey(t)

}
