% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boot3150.R
\name{print.boot3150}
\alias{print.boot3150}
\title{Prints a boot3150 object}
\usage{
\method{print}{boot3150}(x)
}
\arguments{
\item{x}{An object of class boot3150}
}
\value{
This function prints the output for a boot3150 object.
}
\description{
When given a boot3150 object, this function ensures it prints correctly.
}
