% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boot3150ci.R
\name{print.boot3150ci}
\alias{print.boot3150ci}
\title{Prints a boot3150ci object}
\usage{
\method{print}{boot3150ci}(x)
}
\arguments{
\item{x}{An object of class boot3150ci}
}
\value{
This function prints the output for a boot3150ci object.
}
\description{
When given a boot3150ci object, this function ensures it prints correctly.
}
