% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.innerboot3150ci.R
\name{print.innerboot3150ci}
\alias{print.innerboot3150ci}
\title{Prints a innerboot3150ci object}
\usage{
\method{print}{innerboot3150ci}(x)
}
\arguments{
\item{x}{An object of class innerboot3150ci}
}
\value{
This function prints the output for a innerboot3150ci object.
}
\description{
When given a innerboot3150ci object, this function ensures it prints correctly.
}
