% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_plots}
\alias{residual_plots}
\title{Creates residual and QQ plots}
\usage{
residual_plots(model, resid_type = "jackknife", smoother = TRUE, se = T)
}
\arguments{
\item{model}{The model for which we would like these plots. This can be of
class "lm" or "glm" with a binomial family.}

\item{resid_type}{The type of residuals that should be used in the plots.
The options are "jackknife", "raw", "standard", and "pearson". Jackknife
(externally studentized) residuals are used by default for "lm" objects.
Standardized Pearson residuals are used for "glm" objects regardless of what
is entered here.}

\item{smoother}{Logical indicating whether a loess smoother should be added
to the residual plot.}

\item{se}{Logical indicating whether the standard error bands should be added
to the smoother.}
}
\value{
This function returns plots. The user must press "enter" or "return"
to view subsequent plots.
}
\description{
This function creates a residual and a QQ plot for a given model.
}
\examples{
mod <- lm(mpg ~ disp, data = mtcars)
residual_plots(mod)

}
