% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{umap_plot}
\alias{umap_plot}
\title{Create a 1D or 2D UMAP Plot}
\usage{
umap_plot(umap_object, data, color_umap = NULL, ...)
}
\arguments{
\item{umap_object}{An object fit by the \code{umap()} function in the \code{umap}
library.}

\item{data}{The dataset that has the variable to be used for \code{color_umap}.}

\item{...}{Extra arguments for the geom_point() function when plotting in 2D
or for the geom_histogram() function when plotting in 1D.}

\item{res}{The variable that should be used to change the color of the points
on the graph.}
}
\value{
This function returns a ggplot object plot. It can be added onto like
any ggplot object.
}
\description{
This function plots the UMAP layouts in either 1D or 2D and can optionally
color them using a variable from a given dataset.
}
\examples{
data(iris)
iris_umap <- iris[,1:4] |> umap(random_state = 1)
umap_plot(iris_umap, data = iris, color_umap = "Species")

}
